+++ lib/Echidna/UI.hs
@@ -35,11 +35,11 @@ import UnliftIO.Concurrent (killThread, forkIO)
 import qualified Data.Text as T
 
 import Echidna.Campaign
-import Echidna.ABI
 import Echidna.Exec
 import Echidna.Solidity
 import Echidna.Test
 import Echidna.Transaction
+import Echidna.Types (TestState(..), Campaign(..), CampaignConf(..), SolTest, GenDict, Tx(Tx), defaultDict, defSeed, defaultCampaign, coverage, src, ppSolCall)
 
 data UIConf = UIConf { _dashboard :: Bool
                      , _maxTime   :: Maybe Int
